% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auc_pred}
\alias{auc_pred}
\title{auc_pred}
\usage{
auc_pred(beta.hat, V, t0, cov.val, nf = 7)
}
\arguments{
\item{beta.hat}{Estimated regression coefficients}

\item{V}{The asymptotic variance returned from the function covariance_cal}

\item{t0}{A vector of time}

\item{cov.val}{A vector where each element represents a covariate value, ordered to correspond to the elements in \code{beta.hat} for each covariate}

\item{nf}{A numeric value of 3 or 7 for the number of polynomials. When nf = 7, the vector of polynomials = c(t^(-2),t^(-1),t^(-.5),log(t),t^(.5),t,t^2). When nf = 3, the vector of polynomials = c(log(t),t, t^2)}
}
\value{
A matrix of confidence intervals for the time-dependent AUC for each cause of failure.
The matrix has three rows and the number of columns equal to the length of \code{t0}. The first row contains the estimated AUC, the second row contains the lower bound of the confidence interval, and the third row contains the upper bound of the confidence interval.
}
\description{
This is to estimate the the time-dependent AUC with confidence interval for each cause of failure.
}
\examples{
\dontrun{

auc_pred(c(1,0.2,0.1,0.3,0.3,0.5),matrix(rep(1,36),nrow=6),c(1,2),c(1,0), nf=3)

} 
}
