% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{covariance_cal_wrapper}
\alias{covariance_cal_wrapper}
\title{Sum a Numeric Vector Using a C Function}
\usage{
covariance_cal_wrapper(a, b, c, d, e, f, g)
}
\arguments{
\item{a}{A vector of regression coefficients.}

\item{b}{A vector of biomarker values.}

\item{c}{The \code{rho01} vector from the dataframe returned by the \code{data_crossingdc} function. When there are no continuous variables, use the \code{Iil} vector from the dataframe returned by the \code{data_crossingdc} function.}

\item{d}{The \code{rho02} vector from the dataframe returned by the \code{data_crossingdc} function. When there are no continuous variables, use the \code{1-Iil} vector from the dataframe returned by the \code{data_crossingdc} function.}

\item{e}{A transposed matrix where each row represents a covariate, including an intercept as the first row. The rows are ordered to correspond with the regression coefficients in \code{a}.}

\item{f}{The \code{i - 1} vector from the dataframe returned by the \code{data_crossingdc} function.}

\item{g}{The \code{l - 1} vector from the dataframe returned by the \code{data_crossingdc} function.}
}
\value{
A list containing:
\itemize{
 \item \code{sigma1}: The matrix \eqn{\Sigma_1}, used in calculating the asymptotic variance.
 \item \code{sigma2}: The matrix \eqn{\Sigma_2}, also used in calculating the asymptotic variance.
 \item \code{v}: The asymptotic variance \eqn{V}.
}
}
\description{
This wrapper provides an R interface to the `covariance_cal` C++ function.
}
