% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_crossingdc}
\alias{data_crossingdc}
\title{data_crossingdc}
\usage{
data_crossingdc(n, vt, vc, vm, vxs, eta, c0, cont_vars, discrete_vars)
}
\arguments{
\item{n}{Total sample size}

\item{vt}{Vector of observed time}

\item{vc}{Vector of censoring indicator}

\item{vm}{vector of biomarker values}

\item{vxs}{Covariate matrix}

\item{eta}{Censoring indicator for each event type}

\item{c0}{A constant value used for the bandwidth h, h = c0*n^(-1/3)}

\item{cont_vars}{A vector of continuous variables in vxs.}

\item{discrete_vars}{A vector of discrete variables in vxs.}
}
\value{
A dataframe where each row corresponds to a case-control pair at each observed event time.
The dataframe has the following components:
\itemize{
  \item \code{i}: Index for cases.
  \item \code{yi}: Observed event time.
  \item \code{Covariates with "i" suffix}: Covariate columns, with names based on the original covariate name and an "i" suffix for case values.
  \item \code{l}: Index for controls.
  \item \code{Iil}: Indicator for comparison of baseline biomarker values (1 if case biomarker > control biomarker, 0 otherwise).
  \item \code{dif}: Euclidean norm among continuous covariates, when available.
  \item \code{rho01} and \code{rho02}: Numeric vectors representing concordant and discordant event values, respectively, when continuous covariates are present.
  \item \code{rhoweight}: Kernel weight vector, calculated with continuous covariates.
}
}
\description{
This is to reshape data for each cause of failure, the result reshaped data is ready for parameter estimation.
}
\examples{
\dontrun{
data_crossingdc(n, Y0,C0,M0,X,1,2, c("age","bmi"),c("gender","DRUG"))

} 
}
