% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo}
\alias{demo}
\title{Demo data}
\format{
## `demo`
A dataframe of 400 rows and 5 columns:
\describe{
  \item{Y}{Observed event time}
  \item{delta}{Censoring indicator. 0 = censored, 1 = event 1, 2 = event 2}
  \item{xd}{A binary covariate with values of 0 and 1}
  \item{xc}{A continuous covariate}
  \item{M}{Baseline biomarker values}
}
}
\usage{
demo
}
\description{
A data used for demo.
}
\keyword{datasets}
