% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAUCCR-package.R
\docType{package}
\name{tAUCCR-package}
\alias{tAUCCR}
\alias{tAUCCR-package}
\title{tAUCCR: Analyzing Heterogeneity in Biomarker Discriminative Performance under Competing Risks}
\description{
This package performs regression analysis to investigate the impact of covariates on the time-dependent AUC under competing risks, based on the manuscript "Analyzing Heterogeneity in Biomarker Discriminative Performance under Competing Risks".
}
\author{
\strong{Maintainer}: Xinran Huang \email{huangxrtc@gmail.com} (\href{https://orcid.org/0000-0002-8126-9735}{ORCID})

}
\keyword{internal}
